+++ src/test/java/domain/user/PlayerTest.java
@@ -72,23 +72,34 @@ class PlayerTest {
     }
 
     @Test
-    @DisplayName("Player가 가진 카드의 점수의 합이 블랙잭이면 true를 반환한다.")
-    void 카드_블랙잭_인지_확인() {
-        CardHand cardHand = new CardHand();
-        cardHand.add(new Card(Symbol.SPADE, CardNumber.ACE));
-        cardHand.add(new Card(Symbol.CLOVER, CardNumber.KING));
-        cardHand.add(new Card(Symbol.CLOVER, CardNumber.KING));
-        Player player = new Player("name", cardHand);
-        assertThat(player.isBlackjack()).isTrue();
-    }
-
-    @Test
     @DisplayName("Player가 가진 카드의 점수의 합이 블랙잭이 아니면 false를 반환한다.")
     void 카드_블랙잭_아닌지_확인() {
         CardHand cardHand = new CardHand();
         cardHand.add(new Card(Symbol.CLOVER, CardNumber.KING));
         cardHand.add(new Card(Symbol.CLOVER, CardNumber.KING));
         Player player = new Player("name", cardHand);
+        assertThat(player.isBlackjack()).isFalse();
+    }
+
+    @DisplayName("점수 계산 블랙잭인 경우")
+    @Test
+    void 블랙잭인_경우() {
+        CardHand cardHand = new CardHand();
+        cardHand.add(new Card(Symbol.DIAMOND, CardNumber.ACE));
+        cardHand.add(new Card(Symbol.SPADE, CardNumber.JACK));
+        Player player = new Player("name", cardHand);
+
+        assertThat(player.isBlackjack()).isTrue();
+    }
+
+    @DisplayName("점수 계산 블랙잭이 아닌 경우")
+    @Test
+    void 블랙잭이_아닌_경우() {
+        CardHand cardHand = new CardHand();
+        cardHand.add(new Card(Symbol.DIAMOND, CardNumber.ACE));
+        cardHand.add(new Card(Symbol.SPADE, CardNumber.NINE));
+        Player player = new Player("name", cardHand);
+
         assertThat(player.isBlackjack()).isFalse();
     }
 }
